Move prepare test from common test to relayer test
